Organize traits under `ui/src/traits`
