+++ crates/ui/src/prelude.rs
@@ -7,12 +7,12 @@ pub use gpui::{
     WindowContext,
 };
 
-pub use crate::clickable::*;
-pub use crate::disableable::*;
-pub use crate::fixed::*;
-pub use crate::selectable::*;
 pub use crate::styles::{rems_from_px, vh, vw, PlatformStyle, StyledTypography, TextSize};
-pub use crate::visible_on_hover::*;
+pub use crate::traits::clickable::*;
+pub use crate::traits::disableable::*;
+pub use crate::traits::fixed::*;
+pub use crate::traits::selectable::*;
+pub use crate::traits::visible_on_hover::*;
 pub use crate::Spacing;
 pub use crate::{h_flex, v_flex};
 pub use crate::{Button, ButtonSize, ButtonStyle, IconButton, SelectableButton};
