+++ crates/ui/src/traits.rs
@@ -1,0 +1,5 @@
+pub mod clickable;
+pub mod disableable;
+pub mod fixed;
+pub mod selectable;
+pub mod visible_on_hover;
